+++ package.json
@@ -3,22 +3,22 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
+    "@mui/icons-material": "^5.15.15",
+    "@mui/material": "^5.15.15",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-icons": "^5.1.0",
+    "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.1",
-    "web-vitals": "^2.1.4",
-    "@emotion/react": "^11.11.4",
-  "@emotion/styled": "^11.11.5",
-  
-  "@mui/icons-material": "^5.15.15",
-  "@mui/material": "^5.15.15",
-  "@types/react": "^18.0.0",
-  "@types/react-dom": "^18.0.0",
-  "react-router-dom": "^6.2.1",
-  "typescript": "^4.6.3"
+    "typescript": "^4.6.3",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
